refactor: 노선 역 등록 및 삭제 API 설계 변경
- 특정 노선에 대한 API이므로 /lines의 하위 URI로 변경함
