--- src/main/java/subway/ui/SectionController.java
@@ -1,36 +1,0 @@
-package subway.ui;
-
-import javax.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import subway.application.SectionService;
-import subway.dto.SectionRequest;
-
-@RestController
-@RequestMapping("/sections")
-public class SectionController {
-
-    private final SectionService sectionService;
-
-    public SectionController(final SectionService sectionService) {
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> addStations(@RequestBody @Valid SectionRequest sectionRequest) {
-        sectionService.addStations(sectionRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @DeleteMapping("/{lineId}/{stationId}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long lineId, @PathVariable Long stationId) {
-        sectionService.deleteStation(lineId, stationId);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-}
